fix(nx-spring-boot): fix creating spring boot libraries with `gradle`
Align with new official [documentation](https://spring.io/guides/gs/multi-module/) on how to properly create spring boot libraries
