+++ packages/nx-spring-boot/src/generators/project/lib/index.ts
@@ -1,5 +1,5 @@
  export { generateBootProject } from './generate-boot-project';
  export { addBuilInfoTask } from './add-build-info-task';
- export { disableBootJarTask } from './disable-boot-jar-task';
+ export { disableBootGradlePlugin } from './disable-boot-gradle-plugin';
  export { removeBootMavenPlugin } from './remove-spring-boot-plugin';
  export { normalizeOptions } from './normalize-options';
